Turn ModelRenderer and VoxelCache into traits
