+++ OpenRA.Mods.Cnc/Traits/Render/WithVoxelBody.cs
@@ -13,10 +13,11 @@ using System;
 using System.Collections.Generic;
 using OpenRA.Graphics;
 using OpenRA.Mods.Common.Graphics;
+using OpenRA.Mods.Common.Traits;
 using OpenRA.Primitives;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.Common.Traits.Render
+namespace OpenRA.Mods.Cnc.Traits.Render
 {
 	[Desc("Also returns a default selection size that is calculated automatically from the voxel dimensions.")]
 	public class WithVoxelBodyInfo : ConditionalTraitInfo, IRenderActorPreviewVoxelsInfo, Requires<RenderVoxelsInfo>
@@ -29,11 +30,11 @@ namespace OpenRA.Mods.Common.Traits.Render
 
 		public override object Create(ActorInitializer init) { return new WithVoxelBody(init.Self, this); }
 
-		public IEnumerable<ModelAnimation> RenderPreviewVoxels(
+		public IEnumerable<ModelAnimation> RenderPreviewVoxels(IModelCache cache,
 			ActorPreviewInitializer init, RenderVoxelsInfo rv, string image, Func<WRot> orientation, int facings, PaletteReference p)
 		{
 			var body = init.Actor.TraitInfo<BodyOrientationInfo>();
-			var model = init.World.ModelCache.GetModelSequence(image, Sequence);
+			var model = cache.GetModelSequence(image, Sequence);
 			yield return new ModelAnimation(model, () => Offset,
 				() => body.QuantizeOrientation(orientation(), facings),
 				() => false, () => 0, ShowShadow);
@@ -51,7 +52,7 @@ namespace OpenRA.Mods.Common.Traits.Render
 			var body = self.Trait<BodyOrientation>();
 			rv = self.Trait<RenderVoxels>();
 
-			var model = self.World.ModelCache.GetModelSequence(rv.Image, info.Sequence);
+			var model = rv.Renderer.ModelCache.GetModelSequence(rv.Image, info.Sequence);
 			modelAnimation = new ModelAnimation(model, () => info.Offset,
 				() => body.QuantizeOrientation(self.Orientation),
 				() => IsTraitDisabled, () => 0, info.ShowShadow);
