+++ OpenRA.Mods.Cnc/Util.cs
@@ -9,6 +9,8 @@
  */
 #endregion
 
+using System;
+
 namespace OpenRA.Mods.Cnc
 {
 	public static class Util
@@ -64,6 +66,250 @@ namespace OpenRA.Mods.Cnc
 				return SpriteFacings[ClassicIndexFacing(facing, steps)];
 
 			return Common.Util.QuantizeFacing(facing, steps);
+		}
+
+		public static float[] IdentityMatrix()
+		{
+			return new float[]
+			{
+				1, 0, 0, 0,
+				0, 1, 0, 0,
+				0, 0, 1, 0,
+				0, 0, 0, 1,
+			};
+		}
+
+		public static float[] ScaleMatrix(float sx, float sy, float sz)
+		{
+			return new float[]
+			{
+				sx, 0, 0, 0,
+				0, sy, 0, 0,
+				0, 0, sz, 0,
+				0, 0, 0, 1,
+			};
+		}
+
+		public static float[] TranslationMatrix(float x, float y, float z)
+		{
+			return new float[]
+			{
+				1, 0, 0, 0,
+				0, 1, 0, 0,
+				0, 0, 1, 0,
+				x, y, z, 1,
+			};
+		}
+
+		public static float[] MatrixMultiply(float[] lhs, float[] rhs)
+		{
+			var mtx = new float[16];
+			for (var i = 0; i < 4; i++)
+				for (var j = 0; j < 4; j++)
+				{
+					mtx[4 * i + j] = 0;
+					for (var k = 0; k < 4; k++)
+						mtx[4 * i + j] += lhs[4 * k + j] * rhs[4 * i + k];
+				}
+
+			return mtx;
+		}
+
+		public static float[] MatrixVectorMultiply(float[] mtx, float[] vec)
+		{
+			var ret = new float[4];
+			for (var j = 0; j < 4; j++)
+			{
+				ret[j] = 0;
+				for (var k = 0; k < 4; k++)
+					ret[j] += mtx[4 * k + j] * vec[k];
+			}
+
+			return ret;
+		}
+
+		public static float[] MatrixInverse(float[] m)
+		{
+			var mtx = new float[16];
+
+			mtx[0] = m[5] * m[10] * m[15] -
+				m[5] * m[11] * m[14] -
+				m[9] * m[6] * m[15] +
+				m[9] * m[7] * m[14] +
+				m[13] * m[6] * m[11] -
+				m[13] * m[7] * m[10];
+
+			mtx[4] = -m[4] * m[10] * m[15] +
+				m[4] * m[11] * m[14] +
+				m[8] * m[6] * m[15] -
+				m[8] * m[7] * m[14] -
+				m[12] * m[6] * m[11] +
+				m[12] * m[7] * m[10];
+
+			mtx[8] = m[4] * m[9] * m[15] -
+				m[4] * m[11] * m[13] -
+				m[8] * m[5] * m[15] +
+				m[8] * m[7] * m[13] +
+				m[12] * m[5] * m[11] -
+				m[12] * m[7] * m[9];
+
+			mtx[12] = -m[4] * m[9] * m[14] +
+				m[4] * m[10] * m[13] +
+				m[8] * m[5] * m[14] -
+				m[8] * m[6] * m[13] -
+				m[12] * m[5] * m[10] +
+				m[12] * m[6] * m[9];
+
+			mtx[1] = -m[1] * m[10] * m[15] +
+				m[1] * m[11] * m[14] +
+				m[9] * m[2] * m[15] -
+				m[9] * m[3] * m[14] -
+				m[13] * m[2] * m[11] +
+				m[13] * m[3] * m[10];
+
+			mtx[5] = m[0] * m[10] * m[15] -
+				m[0] * m[11] * m[14] -
+				m[8] * m[2] * m[15] +
+				m[8] * m[3] * m[14] +
+				m[12] * m[2] * m[11] -
+				m[12] * m[3] * m[10];
+
+			mtx[9] = -m[0] * m[9] * m[15] +
+				m[0] * m[11] * m[13] +
+				m[8] * m[1] * m[15] -
+				m[8] * m[3] * m[13] -
+				m[12] * m[1] * m[11] +
+				m[12] * m[3] * m[9];
+
+			mtx[13] = m[0] * m[9] * m[14] -
+				m[0] * m[10] * m[13] -
+				m[8] * m[1] * m[14] +
+				m[8] * m[2] * m[13] +
+				m[12] * m[1] * m[10] -
+				m[12] * m[2] * m[9];
+
+			mtx[2] = m[1] * m[6] * m[15] -
+				m[1] * m[7] * m[14] -
+				m[5] * m[2] * m[15] +
+				m[5] * m[3] * m[14] +
+				m[13] * m[2] * m[7] -
+				m[13] * m[3] * m[6];
+
+			mtx[6] = -m[0] * m[6] * m[15] +
+				m[0] * m[7] * m[14] +
+				m[4] * m[2] * m[15] -
+				m[4] * m[3] * m[14] -
+				m[12] * m[2] * m[7] +
+				m[12] * m[3] * m[6];
+
+			mtx[10] = m[0] * m[5] * m[15] -
+				m[0] * m[7] * m[13] -
+				m[4] * m[1] * m[15] +
+				m[4] * m[3] * m[13] +
+				m[12] * m[1] * m[7] -
+				m[12] * m[3] * m[5];
+
+			mtx[14] = -m[0] * m[5] * m[14] +
+				m[0] * m[6] * m[13] +
+				m[4] * m[1] * m[14] -
+				m[4] * m[2] * m[13] -
+				m[12] * m[1] * m[6] +
+				m[12] * m[2] * m[5];
+
+			mtx[3] = -m[1] * m[6] * m[11] +
+				m[1] * m[7] * m[10] +
+				m[5] * m[2] * m[11] -
+				m[5] * m[3] * m[10] -
+				m[9] * m[2] * m[7] +
+				m[9] * m[3] * m[6];
+
+			mtx[7] = m[0] * m[6] * m[11] -
+				m[0] * m[7] * m[10] -
+				m[4] * m[2] * m[11] +
+				m[4] * m[3] * m[10] +
+				m[8] * m[2] * m[7] -
+				m[8] * m[3] * m[6];
+
+			mtx[11] = -m[0] * m[5] * m[11] +
+				m[0] * m[7] * m[9] +
+				m[4] * m[1] * m[11] -
+				m[4] * m[3] * m[9] -
+				m[8] * m[1] * m[7] +
+				m[8] * m[3] * m[5];
+
+			mtx[15] = m[0] * m[5] * m[10] -
+				m[0] * m[6] * m[9] -
+				m[4] * m[1] * m[10] +
+				m[4] * m[2] * m[9] +
+				m[8] * m[1] * m[6] -
+				m[8] * m[2] * m[5];
+
+			var det = m[0] * mtx[0] + m[1] * mtx[4] + m[2] * mtx[8] + m[3] * mtx[12];
+			if (det == 0)
+				return null;
+
+			for (var i = 0; i < 16; i++)
+				mtx[i] *= 1 / det;
+
+			return mtx;
+		}
+
+		public static float[] MakeFloatMatrix(Int32Matrix4x4 imtx)
+		{
+			var multipler = 1f / imtx.M44;
+			return new[]
+			{
+				imtx.M11 * multipler,
+				imtx.M12 * multipler,
+				imtx.M13 * multipler,
+				imtx.M14 * multipler,
+
+				imtx.M21 * multipler,
+				imtx.M22 * multipler,
+				imtx.M23 * multipler,
+				imtx.M24 * multipler,
+
+				imtx.M31 * multipler,
+				imtx.M32 * multipler,
+				imtx.M33 * multipler,
+				imtx.M34 * multipler,
+
+				imtx.M41 * multipler,
+				imtx.M42 * multipler,
+				imtx.M43 * multipler,
+				imtx.M44 * multipler,
+			};
+		}
+
+		public static float[] MatrixAABBMultiply(float[] mtx, float[] bounds)
+		{
+			// Corner offsets.
+			var ix = new uint[] { 0, 0, 0, 0, 3, 3, 3, 3 };
+			var iy = new uint[] { 1, 1, 4, 4, 1, 1, 4, 4 };
+			var iz = new uint[] { 2, 5, 2, 5, 2, 5, 2, 5 };
+
+			// Vectors to opposing corner.
+			var ret = new[]
+			{
+				float.MaxValue, float.MaxValue, float.MaxValue,
+				float.MinValue, float.MinValue, float.MinValue
+			};
+
+			// Transform vectors and find new bounding box.
+			for (var i = 0; i < 8; i++)
+			{
+				var vec = new[] { bounds[ix[i]], bounds[iy[i]], bounds[iz[i]], 1 };
+				var tvec = MatrixVectorMultiply(mtx, vec);
+
+				ret[0] = Math.Min(ret[0], tvec[0] / tvec[3]);
+				ret[1] = Math.Min(ret[1], tvec[1] / tvec[3]);
+				ret[2] = Math.Min(ret[2], tvec[2] / tvec[3]);
+				ret[3] = Math.Max(ret[3], tvec[0] / tvec[3]);
+				ret[4] = Math.Max(ret[4], tvec[1] / tvec[3]);
+				ret[5] = Math.Max(ret[5], tvec[2] / tvec[3]);
+			}
+
+			return ret;
 		}
 	}
 }
