+++ mods/cnc/mod.yaml
@@ -240,8 +240,6 @@ TerrainFormat: DefaultTerrain
 
 SpriteSequenceFormat: ClassicTilesetSpecificSpriteSequence
 
-ModelSequenceFormat: PlaceholderModelSequence
-
 AssetBrowser:
 	SpriteExtensions: .shp, .tem, .win, .sno, .des, .jun
 	AudioExtensions: .aud, .wav, .v00, .v01, .v02, .v03, .var
