+++ mods/ts/mod.yaml
@@ -273,8 +273,6 @@ TerrainFormat: DefaultTerrain
 
 SpriteSequenceFormat: TilesetSpecificSpriteSequence
 
-ModelSequenceFormat: VoxelModelSequence
-
 AssetBrowser:
 	SpriteExtensions: .shp, .tem, .sno
 	ModelExtensions: .vxl
